Refactor decorator into pipeline.
