+++ consultation_analyser/pipeline/decorators.py
