+++ consultation_analyser/pipeline/llm_summariser.py
@@ -14,8 +14,8 @@ from langchain.prompts import PromptTemplate
 from langchain.pydantic_v1 import BaseModel, ValidationError
 from langchain_core.exceptions import OutputParserException
 
-from consultation_analyser.consultations.decorators.sagemaker_endpoint_status_check import check_and_launch_sagemaker
 from consultation_analyser.consultations.models import Answer, Theme
+from consultation_analyser.pipeline.decorators import check_and_launch_sagemaker
 
 MODEL_ENCODING = tiktoken.get_encoding(
     "cl100k_base"
